增加上传行驶证，实时监测界面（等接口）

+++ BusHelp/SettingViewController.m
@@ -228,7 +228,7 @@
             _orgArray = orgArray;
             _org = [_orgArray firstObject];
             NSLog(@"组织名称：%@",_org.name);
-            NSLog(@"status:%i",_org.status.integerValue);
+            NSLog(@"status:%li",(long)_org.status.integerValue);
             if (_org != nil ) {
                 showApplyCell=YES;
             }else
